Remove k8s_yaml_manager.py file
